requirements.txt
appnope==0.1.3
asttokens==2.1.0
attrs==22.1.0
backcall==0.2.0
boto3==1.26.8
botocore==1.29.8
click==8.1.3
contextlib2==21.6.0
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
dill==0.3.6
dnspython==2.2.1
entrypoints==0.4
executing==1.2.0
fonttools==4.38.0
google-pasta==0.2.0
importlib-metadata==4.13.0
ipykernel==6.17.1
ipython==8.6.0
jedi==0.18.1
jmespath==1.0.1
jupyter-client==7.4.5
jupyter-core==5.0.0
kiwisolver==1.4.4
matplotlib==3.6.2
matplotlib-inline==0.1.6
multiprocess==0.70.14
nest-asyncio==1.5.6
numpy==1.23.4
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathos==0.3.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.3
pox==0.3.2
ppft==1.7.6.6
prompt-toolkit==3.0.32
protobuf==3.20.3
protobuf3-to-dict==0.1.5
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
PyExifTool==0.5.5
Pygments==2.13.0
pymongo==4.3.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
pyzmq==24.0.1
s3transfer==0.6.0
sagemaker==2.116.0
schema==0.7.5
six==1.16.0
smdebug-rulesconfig==1.0.1
stack-data==0.6.0
tornado==6.2
tqdm==4.65.0
traitlets==5.5.0
urllib3==1.26.12
wcwidth==0.2.5
zipp==3.10.0
